refactor(player): coalesce seek repaint (L6); share PlayerEngineSurface (M2)
- L6: refreshTimeUISoon now cancels its prior follow-up repaints before
  scheduling fresh ones (cancellable DispatchWorkItems) — a rapid scrub
  burst no longer stacks dozens of pending closures. Same two-delay
  (0.15s/0.4s) post-seek repaint behavior, just bounded.
- M2: extracted the byte-equivalent EngineSurface / OfflineEngineSurface
  into one shared PlayerEngineSurface (iOS PiPVideoView + tvOS VideoView
  split preserved). Both VLC presenters now use it.

Behavior-preserving; iOS+tvOS build + 40 tests green. Runtime scrub/PiP
behavior still warrants an on-device pass. xcodegen regenerated.
